+++ docs/code/pages.py
@@ -1,0 +1,14 @@
+references = [
+    "chemistry/H",
+    "chemistry/He",
+    "chemistry/Li",
+    "chemistry/Be",
+]
+
+
+titled_references = {
+    "Element 1": "chemistry/H",
+    "Element 2": "chemistry/He",
+    "Element 3": "chemistry/Li",
+    "Element 4": "chemistry/Be",
+}