+++ source/added_value/sphinxext.py
@@ -3,6 +3,7 @@ from added_value.any_items_role import any_items_role
 from added_value.format_role import format_role
 from added_value.items_list_directive import ItemsListDirective
 from added_value.items_table_directive import ItemsTableDirective
+from added_value.items_toc_directive import ItemsTableOfContentsDirective
 from added_value.str_role import str_role
 from added_value.repr_role import repr_role
 from added_value.version import __version__
@@ -19,5 +20,6 @@ def setup(app):
     app.add_role("any-items", any_items_role)
     app.add_directive("items-table", ItemsTableDirective)
     app.add_directive("items-list", ItemsListDirective)
+    app.add_directive("items-toc", ItemsTableOfContentsDirective)
 
     return {"version": __version__}